Add ORM server
Creates basic routes for saving and reading a Challenge
