+++ docker-compose.yml
@@ -42,7 +42,20 @@ services:
         POSTGRES_DB: mart
     volumes:
       - postgres:/var/lib/postgresql/data 
+    networks:
+      - database-net
+  orm:
+    ports:
+      - "8084:8080"
+    build:
+      context: ./
+      dockerfile: orm/Dockerfile
+    networks:
+      - maratona-net
+      - database-net
+
 networks: 
   maratona-net: 
+  database-net:
 volumes:
-  postgres:
+    postgres:
